👌🏻 Move bootstrapRefresVersions() to de.fayard.versions
